+++ Net.Pokeshot.JiveSdk/Models/Dto/JiveDEADestination.cs
@@ -16,8 +16,8 @@ namespace Net.Pokeshot.JiveSdk.Models.Dto
         public string name {get;set;}
         public string displayName {get;set;}
         public string description {get;set;}
-        public int creationDate {get;set;}
-        public int modificationDate {get;set;}
+        public long creationDate {get;set;}
+        public long modificationDate {get;set;}
         public string url {get;set;}
         public string status {get;set;}
         public bool isVisibleToPartner {get;set;}
